chore(yaml): split yaml into DockerComposeYaml and KubernetesYaml
